+++ src/FreeElephants/TSxUnit/Exception/InvalidArgumentException.ts
@@ -1,0 +1,7 @@
+///<reference path="BaseException.ts"/>
+
+namespace FreeElephants.TSxUnit.Exception {
+
+    export class InvalidArgumentException extends BaseException {
+    }
+}